Multiple simultaneous 'make test's can fail because of port in use
From: Joe McMahon
(https://rt.cpan.org/Public/Bug/Display.html?id=54322)

t/01-epoll.t uses fixed port numbers; this can cause the test to fail if
multiple copies are running on the same server at the same time. This is
a not infrequent occurrence on our continuous build server, and leads to
pings for failing builds when nothing is actually wrong.

The attached patch uses IO::Socket::INET's built-in port randomization
to ensure we don't have a test failure because of a busy socket.